Move profile check to custom kernel manager and update prompt on restart
As the spock profile can change while the frontend is running, it is
necessary to check the profile on each start of the kernel and update
the banner and the prompt to always contain the correct spock version
and door name, respectively.

Using a custom kernel manager ensures that the spock profile is checked
always before the kernel is started, e.g., by the auto-restarter or by
calls to QtSpock.restart_kernel. From the check of the profile until it
is actually loaded by the kernel, a small time window for unchecked
changes remains.